Refactor seeding
