+++ src/client/java/teammates/client/scripts/sql/SeedNonCourseData.java
@@ -1,6 +1,5 @@
 package teammates.client.scripts.sql;
 
-import java.io.IOException;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
@@ -11,74 +10,20 @@ import java.util.Random;
 import java.util.Set;
 import java.util.UUID;
 
-import com.google.cloud.datastore.DatastoreOptions;
-import com.googlecode.objectify.ObjectifyFactory;
-import com.googlecode.objectify.ObjectifyService;
-import com.googlecode.objectify.util.Closeable;
-
-import teammates.common.util.Config;
 import teammates.client.connector.DatastoreClient;
 import teammates.client.scripts.GenerateUsageStatisticsObjects;
-import teammates.common.datatransfer.DataBundle;
 import teammates.common.datatransfer.NotificationStyle;
 import teammates.common.datatransfer.NotificationTargetUser;
 import teammates.common.datatransfer.attributes.AccountRequestAttributes;
-import teammates.test.FileHelper;
-import teammates.common.util.JsonUtils;
-import teammates.logic.api.LogicExtension;
-import teammates.logic.core.LogicStarter;
-import teammates.storage.api.OfyHelper;
 import teammates.storage.entity.Account;
 import teammates.storage.entity.AccountRequest;
 import teammates.storage.entity.Notification;
 
-public class SeedDb extends DatastoreClient {
-    private final LogicExtension logic = new LogicExtension();
-
-    private Closeable closeable;
-
-    public void setupDbLayer() throws Exception {
-        LogicStarter.initializeDependencies();
-    }
-
-    public void setupObjectify() {
-        DatastoreOptions.Builder builder = DatastoreOptions.newBuilder().setProjectId(Config.APP_ID);
-        ObjectifyService.init(new ObjectifyFactory(builder.build().getService()));
-        OfyHelper.registerEntityClasses();
-
-        closeable = ObjectifyService.begin();
-    }
-
-    public void tearDownObjectify() {
-        closeable.close();
-    }
-
-    protected String getSrcFolder() {
-        return "src/client/java/teammates/client/scripts/sql/";
-    }
-
-    protected DataBundle loadDataBundle(String jsonFileName) {
-        try {
-            String pathToJsonFile = getSrcFolder() + jsonFileName;
-            String jsonString = FileHelper.readFile(pathToJsonFile);
-            return JsonUtils.fromJson(jsonString, DataBundle.class);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    protected DataBundle getTypicalDataBundle() {
-        return loadDataBundle("typicalDataBundle.json");
-    }
-
-    protected Instant getRandomInstant() {
-        return Instant.now();
-    }
-
+public class SeedNonCourseData extends DatastoreClient {
     protected void persistAdditionalData() {
         int ENTITY_SIZE = 10000;
         // Each account will have this amount of read notifications
-        int READ_NOTIFICATION_SIZE = 5;
+        int READ_NOTIFICATION_SIZE = 1;
         int NOTIFICATION_SIZE = 1000;
         assert (NOTIFICATION_SIZE >= READ_NOTIFICATION_SIZE);
 
@@ -97,21 +42,22 @@ public class SeedDb extends DatastoreClient {
             }
             notificationUUIDs.add(notificationUUID.toString());
             notificationsUUIDSeen.add(notificationUUID.toString());
+
             // Since we are not using logic class, referencing
             // MarkNotificationAsReadAction.class and CreateNotificationAction.class
             // endTime is to nearest milli not nanosecond
-            Instant endTime = getRandomInstant().truncatedTo(ChronoUnit.MILLIS);
+            Instant endTime = Instant.now().truncatedTo(ChronoUnit.MILLIS);
             Notification notification = new Notification(
                     notificationUUID.toString(),
-                    getRandomInstant(),
+                    Instant.now(),
                     endTime,
                     NotificationStyle.PRIMARY,
                     NotificationTargetUser.INSTRUCTOR,
                     notificationUUID.toString(),
                     notificationUUID.toString(),
                     false,
-                    getRandomInstant(),
-                    getRandomInstant());
+                    Instant.now(),
+                    Instant.now());
             try {
                 ofy().save().entities(notification).now();
                 notificationEndTimes.put(notificationUUID.toString(), notification.getEndTime());
@@ -165,27 +111,23 @@ public class SeedDb extends DatastoreClient {
     }
 
     protected void persistData() {
-        // Persisting basic data bundle
-        DataBundle dataBundle = getTypicalDataBundle();
         try {
-            logic.persistDataBundle(dataBundle);
             persistAdditionalData();
         } catch (Exception e) {
-            e.printStackTrace();
+            log(e.toString());
         }
     }
 
     public static void main(String[] args) throws Exception {
-        new SeedDb().doOperationRemotely();
+        new SeedNonCourseData().doOperationRemotely();
     }
 
     @Override
     protected void doOperation() {
         try {
-            // LogicStarter.initializeDependencies();
             this.persistData();
         } catch (Exception e) {
-
+            log(e.toString());
         }
     }
 }
